+++ examples/triangle_glsl.py
@@ -45,17 +45,17 @@ void main()
 # %% The wgpu calls
 
 
-def setup_drawing_sync(canvas, power_preference="high-performance", limits=None):
-    """Regular function to set up a viz on the given canvas."""
+def setup_drawing_sync(context, power_preference="high-performance", limits=None):
+    """Regular function to set up a viz on the given context."""
 
     adapter = wgpu.gpu.request_adapter_sync(power_preference=power_preference)
     device = adapter.request_device_sync(required_limits=limits)
 
-    render_pipeline = get_render_pipeline(canvas, device)
-    return get_draw_function(canvas, device, render_pipeline)
+    render_pipeline = get_render_pipeline(context, device)
+    return get_draw_function(context, device, render_pipeline)
 
 
-def get_render_pipeline(canvas, device):
+def get_render_pipeline(context, device):
     vert_shader = device.create_shader_module(label="triangle_vert", code=vertex_shader)
     frag_shader = device.create_shader_module(
         label="triangle_frag", code=fragment_shader
@@ -64,9 +64,8 @@ def get_render_pipeline(canvas, device):
     # No bind group and layout, we should not create empty ones.
     pipeline_layout = device.create_pipeline_layout(bind_group_layouts=[])
 
-    present_context = canvas.get_context("wgpu")
-    render_texture_format = present_context.get_preferred_format(device.adapter)
-    present_context.configure(device=device, format=render_texture_format)
+    render_texture_format = context.get_preferred_format(device.adapter)
+    context.configure(device=device, format=render_texture_format)
 
     return device.create_render_pipeline(
         layout=pipeline_layout,
@@ -97,9 +96,9 @@ def get_render_pipeline(canvas, device):
     )
 
 
-def get_draw_function(canvas, device, render_pipeline):
+def get_draw_function(context, device, render_pipeline):
     def draw_frame():
-        current_texture = canvas.get_context("wgpu").get_current_texture()
+        current_texture = context.get_current_texture()
         command_encoder = device.create_command_encoder()
 
         render_pass = command_encoder.begin_render_pass(
@@ -127,6 +126,8 @@ if __name__ == "__main__":
     from rendercanvas.auto import RenderCanvas, loop
 
     canvas = RenderCanvas(size=(640, 480), title="wgpu triangle glsl example")
-    draw_frame = setup_drawing_sync(canvas)
+    context = canvas.get_wgpu_context()
+
+    draw_frame = setup_drawing_sync(context)
     canvas.request_draw(draw_frame)
     loop.run()