+++ tests/renderutils.py
@@ -276,7 +276,7 @@ def render_to_screen(
         },
     )
 
-    present_context = canvas.get_context("wgpu")
+    present_context = canvas.get_wgpu_context()
     present_context.configure(device=device, format=None)
 
     def draw_frame():